feat: #8 習慣カードに連続日数バッジを追加、calcStats/calcCurrentStreakをutils/stats.jsに共通化
