+++ src/components/HabitButton.css
@@ -74,3 +74,13 @@
   font-weight: 700;
   opacity: 0.9;
 }
+
+.habit-streak {
+  position: absolute;
+  bottom: 5px;
+  right: 7px;
+  font-size: 0.6rem;
+  font-weight: 700;
+  opacity: 0.7;
+  line-height: 1;
+}
